rename suggestions file
